requirements.txt
 django==5.2.4
 djangorestframework==3.15.2
 djangorestframework-simplejwt==5.4.0
 django-filter==2.4.0
 drf-spectacular==0.26.3
 django-cors-headers==4.5.0
 psycopg2-binary==2.9.10
 django-redis==5.4.0
 python-dotenv